requirements/requirements.txt
amqp==5.2.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron_descriptor==1.4.0
cryptography==41.0.4
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.4
django-allauth==0.55.2
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-render-partial==0.4
django-social-share==2.3.0
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
executing==2.0.1
ghasedakpack==0.1.13
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.17.2
jedi==0.19.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.4
Markdown==3.4.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==1.7.1
pytest==7.4.4
pytest-django==4.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.12
